requirements.txt
# Core ML dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0,<2.0
Pillow>=8.3.0

# Data processing and computer vision
opencv-python>=4.5.0
scikit-image>=0.18.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# Web frameworks for demo
gradio>=3.0.0
streamlit>=1.0.0
fastapi>=0.95
uvicorn[standard]>=0.21

# Scientific computing
scipy>=1.7.0
pandas>=1.3.0

# Utilities and configuration
ninja>=1.10.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0
pyspng>=0.1.0

# CycleGAN dependencies
dominate>=2.6.0
visdom>=0.1.8

# Evaluation metrics
pytorch-fid>=0.3.0
lpips>=0.1.4

# Data augmentation
albumentations>=1.0.0

# Development utilities
click>=8.0.0
pyyaml>=5.4.0
tensorboard>=2.7.0
einops>=0.3.0

# Additional dependencies
pillow>=9.0
python-multipart>=0.0.6
prometheus_client>=0.16.0
python-json-logger>=2.0
